Function phile::dalgen::generate_dal

pub fn generate_dal(
    sqir: &Sqir, 
    params: &CodegenParams, 
    wp: &mut WriterProvider
) -> Result<()>

Given the SQIR representation of a program, and some configuration parameters, generates a Database Abstraction Layer and writes the code into io::Writes.

Arguments:

sqir: Optimized SQIR intermediate representation for a PHiLe program.

params: code generation parameters; see the docs for CodegenParams.

wp: a provider of io::Writes that will accumulate generated code.

Return value:

Ok(()), if code generation completed successfully.

Err(Error), if an error occurred. This is typically an I/O error (Error::IO) or an Error::Semantic induced by some special requirement of the backend (e.g. a missing namespace when generating Go code).