Enum phile::sqir::Value

pub enum Value {
    Placeholder,
    Nil,
    BoolConst(bool),
    IntConst(u64),
    FloatConst(f64),
    StringConst(String),
    Function(Function),
    Call(Call),
    FuncArg {
        func: WkExpr,
        index: usize,
    },
    Load(WkExpr),
    OptionalWrap(RcExpr),
    Ignore(RcExpr),
    Neg(RcExpr),
    Add(RcExpr, RcExpr),
    Sub(RcExpr, RcExpr),
    Mul(RcExpr, RcExpr),
    Div(RcExpr, RcExpr),
    Mod(RcExpr, RcExpr),
    Eq(RcExpr, RcExpr),
    Neq(RcExpr, RcExpr),
    Lt(RcExpr, RcExpr),
    LtEq(RcExpr, RcExpr),
    Gt(RcExpr, RcExpr),
    GtEq(RcExpr, RcExpr),
    And(RcExpr, RcExpr),
    Or(RcExpr, RcExpr),
    Not(RcExpr),
    Intersect(RcExpr, RcExpr),
    Union(RcExpr, RcExpr),
    SetDiff(RcExpr, RcExpr),
    Branch(Branch),
    Seq(Vec<RcExpr>),
    Array(Vec<RcExpr>),
    Tuple(Vec<RcExpr>),
    StructLiteral(BTreeMap<String, RcExpr>),
    Map(Map),
    Reduce(Reduce),
    Filter(Filter),
    Sort(Sort),
    Group(Group),
    Join,
    Insert,
    Update,
    Project(AttrRef),
    GroupBy(AttrRef),
    SortBy {
        entity: RcType,
        fields: Vec<(String, SortDirection)>,
    },
    Sum(AttrRef),
    Avg(AttrRef),
    Min(AttrRef),
    Max(AttrRef),
}

Represents the value of an expression node.

There are two kinds of values. One kind belongs in the Core: only these variants are emitted directly by SqirGen. They describe database operations as generally and as uniformly as possible.

The other kind of value variants describe simpler, specialized, frequently-encountered database operations. These are only emitted by SqirOpt, as the result of collapsing more general SQIR, so that the resulting code is easier to map to actualy database queries efficiently.

Variants
Placeholder

Value of a forward-declared function

Nil

A nil constant.

BoolConst(bool)

A Boolean constant.

IntConst(u64)

An integral constant.

FloatConst(f64)

A floating-point constant.

StringConst(String)

A string constant.

Function(Function)

A function definition (~lambda expression).

Call(Call)

A function call (~application).

FuncArg

A function argument. This has its own variant because it makes optimizations significantly easier.

Fields of FuncArg

func: WkExpr

A back-pointer to the containing function.

index: usize

0-based index of the argument within its containing function.

Load(WkExpr)

A name reference (~substitution). The wrapped expression is the underlying expression referred to by the "variable". Variable declarations are implicit using named expressions. It must be weak, because the body of a named (eg. global) recursive function can refer to itself.

OptionalWrap(RcExpr)

T -> T? conversion. Usually emitted from an implicit promotion. Each conversion generates a new temporary, therefore these conversions cannot form a reference cycle, so it's OK to make their associated value a strong pointer.

Ignore(RcExpr)

Explicit T -> () conversion. Usually emitted by an expression statement.

Neg(RcExpr)

Numeric unary negation operation.

Add(RcExpr, RcExpr)

Addition operation.

Sub(RcExpr, RcExpr)

Subtraction operation.

Mul(RcExpr, RcExpr)

Multiplication operation.

Div(RcExpr, RcExpr)

Division operation.

Mod(RcExpr, RcExpr)

Modulus (or remainder?) operation.

Eq(RcExpr, RcExpr)

Equality comparison.

Neq(RcExpr, RcExpr)

Inequality comparison.

Lt(RcExpr, RcExpr)

"Less than" comparison.

LtEq(RcExpr, RcExpr)

"Less than or equal to" comparison.

Gt(RcExpr, RcExpr)

"Greater than" comparison.

GtEq(RcExpr, RcExpr)

"Greater than or equal to" comparison.

And(RcExpr, RcExpr)

Logical conjunction.

Or(RcExpr, RcExpr)

Logical disjunction.

Not(RcExpr)

Logical negation.

Intersect(RcExpr, RcExpr)

Set intersection.

Union(RcExpr, RcExpr)

Set union.

SetDiff(RcExpr, RcExpr)

Set difference.

Branch(Branch)

Multi-way branch.

Seq(Vec<RcExpr>)

An ordered sequence of expressions; a block. Evaluates each expression and yields the last one.

Array(Vec<RcExpr>)

An array literal.

Tuple(Vec<RcExpr>)

A tuple literal.

StructLiteral(BTreeMap<String, RcExpr>)

A struct literal. Keys are field names, values are the corresponding expressions.

Map(Map)

Generalized projection.

Reduce(Reduce)

Generalized aggregation.

Filter(Filter)

Generalized selection.

Sort(Sort)

Generalized sorting.

Group(Group)

Generalized grouping.

Join

A join operation between entities. TODO(H2CO3): design + implement this.

Insert

An insertion into some entities in the database. TODO(H2CO3): design + implement this.

Update

An update affecting some entities in the database. TODO(H2CO3): design + implement this.

Project(AttrRef)

Simple projection of one or more named columns of the given type. Only emitted by SQIROpt.

GroupBy(AttrRef)

Simple grouping based on one or more named columns of the given type. Column names are in decreasing order of significance ("outer first"). Only emitted by SQIROpt.

SortBy

Simple sorting based on one or more named columns of the given type. Column names are in decreasing order of (lexicographical) significance. Only emitted by SQIROpt.

Fields of SortBy

entity: RcType

The entity to sort.

fields: Vec<(String, SortDirection)>

Names of fields used as sort keys, in lexicographical order of importance.

Sum(AttrRef)

Project the sum of some attributes of a given type. Only emitted by SQIROpt.

Avg(AttrRef)

Project the average of some attributes of a given type. Only emitted by SQIROpt.

Min(AttrRef)

Project the minimum of some attributes of a given type. Only emitted by SQIROpt.

Max(AttrRef)

Project the maximum of some attributes of a given type. Only emitted by SQIROpt.