Struct phile::sqir::Sqir

pub struct Sqir {
    pub named_types: BTreeMap<String, RcType>,
    pub decimal_types: HashMap<(usize, usize), RcType>,
    pub optional_types: HashMap<RcType, RcType>,
    pub pointer_types: HashMap<RcType, RcType>,
    pub array_types: HashMap<RcType, RcType>,
    pub tuple_types: HashMap<Vec<RcType>, RcType>,
    pub function_types: HashMap<(Vec<RcType>, RcType), RcType>,
    pub relations: HashMap<(RcType, String), Relation>,
    pub globals: BTreeMap<Option<String>, BTreeMap<String, RcExpr>>,
}

Top-level type for wrapping SQIR for a complete program. Field names are (hopefully) self-explanatory.

*_types fields are type caches: in LLVM's style, types are interned and structurally uniqued, so that comparing them for equality is a simple pointer comparison. This requires them to be cached. Cache keys are chosen so that they uniquely identify the structure of the type. For named types, this is the name of the type; for wrapper types (e.g. optionals), this is just the wrapped type, for tuples, it is the product of the contained types, for decimals, it's a pair of the number of integral and fractional bits, etc.

relations contains all existing relations between entities; keys are (referring entity, referring field name) pairs.

globals stores all global values (e.g. top-level functions); keys are namespaces/impl names: None for no namespace, Some("Name") for impl Name.

Fields
named_types: BTreeMap<String, RcType> decimal_types: HashMap<(usize, usize), RcType> optional_types: HashMap<RcType, RcType> pointer_types: HashMap<RcType, RcType> array_types: HashMap<RcType, RcType> tuple_types: HashMap<Vec<RcType>, RcType> function_types: HashMap<(Vec<RcType>, RcType), RcType> relations: HashMap<(RcType, String), Relation> globals: BTreeMap<Option<String>, BTreeMap<String, RcExpr>>
Methods

impl Sqir

fn new() -> Sqir

Creates an "empty" SQIR value, populated with builtins only.

fn unique_relations(&self) -> BTreeSet<&Relation>

Relations can be stored either once or twice in the relations field (because they need to be associated with named sides for easy querygen), but schemagen still requires a non-redundant set of relations. By convention, relations contains relations in a way that the RelationSide corresponding to the key appears on the LHS, whereas the referred RelationSide is the RHS.