Function phile::util::grapheme_count_by

pub fn grapheme_count_by<P: Fn(&str) -> bool>(string: &str, pred: P) -> usize

Counts the grapheme clusters in a string that satisfy a condition.

Arguments:

string: a string slice.

pred: a predicate function invoked for each extended grapheme cluster in string.

Return value:

The number of extended grapheme clusters in string for which pred returned true.