Struct phile::util::RcCell

pub struct RcCell<T: ?Sized> { /* fields omitted */ }

A reference counted, dynamically borrow checked smart pointer. Like Rc<RefCell<T>>, but with a more convenient interface.

Methods

impl<T> RcCell<T>

fn new(value: T) -> RcCell<T>

Creates an RcCell owning the value.

fn borrow(&self) -> Result<Ref<T>>

Attempts to immutably borrow the pointed value.

Return value:

Ok(Ref<T>), if an immutable borrow is dynamically possible.

Err(Error::Borrow), if an immutable borrow is dynamically impossible, e.g. because there is outstanding mutable loan.

fn borrow_mut(&self) -> Result<RefMut<T>>

Attempts to mutably borrow the pointed value.

Return value:

Ok(Ref<T>), if a mutable borrow is dynamically possible.

Err(Error::BorrowMut), if a mutable borrow is dynamically impossible, e.g. because there is outstanding immutable loan.

fn as_weak(&self) -> WkCell<T>

Converts the strong pointer to a weak pointer.

fn strong_count(&self) -> usize

Returns the number of strong pointers pointing to the inner value.

fn weak_count(&self) -> usize

Returns the number of weak pointers pointing to the inner value.